massif/tests/pages_as_heap.c: Fix a compiler warning
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@13318

+++ massif/tests/pages_as_heap.c
@@ -7,7 +7,7 @@ int main () {
   int i;
   int inc_dec;
   int delta;
-  int brk_stat;
+  intptr_t brk_stat;
 
   // loop to first increase, then decrease
   for (inc_dec = 1; inc_dec >= -1; inc_dec-=2) {
@@ -16,7 +16,7 @@ int main () {
         if (0) printf("initial brk value for inc_dec %d delta %d: %p\n",
                inc_dec, delta, sbrk(0));
         for (i=0; i<MAX; i++) {
-           brk_stat = brk(sbrk(0) + inc_dec * delta);
+           brk_stat = (intptr_t)brk(sbrk(0) + inc_dec * delta);
            if (brk_stat == -1) {
               printf("brk value at failure: %p\n", sbrk(0));
               perror ("brk() failed!\n");
